@@ -1,5 +1,5 @@
 // tailwind.config.js
-module.exports = {
+export default {
   darkMode: 'class', // Not 'media'
   content: [
     './app/**/*.{js,ts,jsx,tsx}',
